Major refactor
I removed the use of environment variables to indicate what we want to migrate and to where. These values are now passed into the migrate function as source and target arguments.

The main function was refactored and renmaed to the migrate function. The function now expects two arguments to be passed, a source and target. All functionality has been updated to use these two new arguments. The function is now also a generator function.

The getProjectData function was added to help abstract getting the project data from a repo or an org. We now try to return the same data structure for both. This lets us simplify our code and helps us manage any unforseen errors.