Remove dotenv and add github client import
I've added the import for the new GiHub client module.

Since we are now using the device flow to obtain the user's access_token I've removed the usage of the dotenv module.
I'm not going to remove the module, it will still be important for local development and testing. I just need
to give some thought as to how I want to use it.