Move graphql path variable to main.ts
I've moved the graphql queries path variable into the main.ts file from the utils.ts.
The main reason for this was to fix a bug when using ncc to distribute the CLI. NCC packages
all code into a single file which is outside of the libs folder. This broke the way we were
building the path to the graphql queries.

Once moved into the main.ts file the path creation was at the same level as the distribution CLI file.