Add settings for NPM publication
I've added the publshConfig and files entries to the package.json. This will help
us publish the CLI correctly.

I've also updated the main and bin entries to point to dist/index.js and not cli.js. This is
because ncc defaults to creating a index.js and for publishing purposes this seems correct to me.