Add issue transfer functionality
The project had an issue when trying to move project boards between repos. When trying to move cards that had issues or
pull request as their content, the app would fail. This is because the content did not exist in the target repos.

@chocrates recommended adding this functionality in issue #17.

There were two option for this.

1. Tansfer the issue into the target repository
2. Create a not with a link to the issue or pull request.

I've implemented the issue transfer by added a new prompt for the use to indicate if they want to transfer the issue.
If they do not I will create a note in the target's column with a link to the issue or pull request.

Pull request can't be transfered.

This commit adds:

- new logic to handle this, along with edits to the graphql queries to get the needed repo ids for the issue
transfer.
- A new query was also added transer-issue.graphql
- A new helper function was added transferIssue
- A new prompt was added as well.